Update index page to get content from Wordpress API
Updating index page with the feature to show data extracted
from the Wordpress API, this data is extracted from
a page that needs to have the name 'content-page-home' because
the page will be searched by its URL Slug.

If the page hasn't been created a message with the instructions
will be displayed. The index page gets and passes it to the
layout component that contains the Header component, after
the data is passed from layout component to Header component.

Important; for this to works it is important install the
Advanced Custom Fields plugin in our WP backend and
create specific fields, this fields names are in the index
page graqhql query.
https://www.wpgraphql.com/acf/

  - Adding validations in gatsby-node file to create pages
    from data of the Wordpress API
  - Updating .eslintrc file whit rule to pass error that
    occurred for the validations added in gatsby-config
  - Updating the index page to get custom fields from the Wordpress
    API
  - Updating the header component to display data from the
    Wordpress API