Add menu component to display menu options in a top navbar
The header component only will be displayed on the main page
of the website.

The menu component has a scroll eventListener, this event is
to set the menu background when the route is "/" because
the menu component won't have background and image when is
in the top of the page, in the other pages, the menu will have
background and image and the scroll eventListener will be
disabled.

  - Updating menu component file
  - Updating layout component to receive showHeader prop
  - Updating the page-template to display the title page